corrigindo a fonte

+++ models/dashbordTVModel.py
@@ -510,7 +510,7 @@ def OutrosFat(ano, empresa):
                                    'DEV MP':faturamento_mes_DEV, 'DEV MP Acumulado':faturamento_acumulado_DEV})
     df_faturamento['total'] = df_faturamento['VD Mostruario'].str.replace('R\$ ', '').str.replace('.', '').str.replace(
         ',', '.')
-    df_faturamento['total'] = df_faturamento['total'].astype(int)
+    #df_faturamento['total'] = df_faturamento['total'].astype(int)
     total = dataframe['faturado'].sum()
     total = "{:,.2f}".format(total)
     total = 'R$ ' + str(total)
